Move rasa_example to harmoni_models
