--- harmoni_dialogues/harmoni_bot/src/rasa_example/data/rules.yml
@@ -1,13 +1,0 @@
-version: "2.0"
-
-rules:
-
-- rule: Say goodbye anytime the user says goodbye
-  steps:
-  - intent: goodbye
-  - action: utter_goodbye
-
-- rule: Say 'I am a bot' anytime the user challenges
-  steps:
-  - intent: bot_challenge
-  - action: utter_iamabot
