--- harmoni_dialogues/harmoni_bot/src/rasa_greeter/credentials.yml
@@ -1,36 +1,0 @@
-# This file contains the credentials for the voice & chat platforms
-# which your bot is using.
-# https://rasa.com/docs/rasa/messaging-and-voice-channels
-
-rest:
-#  # you don't need to provide anything here - this channel doesn't
-#  # require any credentials
-
-addons.custom_channel.MyIO:
-  username: "user_name"
-  another_parameter: "some value"
-
-#facebook:
-#  verify: "<verify>"
-#  secret: "<your secret>"
-#  page-access-token: "<your page access token>"
-
-#slack:
-#  slack_token: "<your slack token>"
-#  slack_channel: "<the slack channel>"
-#  proxy: "<your HTTP outgoing proxy>"
-
-#socketio:
-#  user_message_evt: <event name for user message>
-#  bot_message_evt: <event name for bot messages>
-#  session_persistence: <true/false>
-
-#mattermost:
-#  url: "https://<mattermost instance>/api/v4"
-#  token: "<bot token>"
-#  webhook_url: "<callback URL>"
-
-# This entry is needed if you are using Rasa X. The entry represents credentials
-# for the Rasa X "channel", i.e. Talk to your bot and Share with guest testers.
-rasa:
-  url: "http://localhost:5002/api"
