--- harmoni_dialogues/harmoni_bot/src/rasa_greeter/data/rules.yml
@@ -1,15 +1,0 @@
-version: "2.0"
-stories:
-- story: counting
-  steps:
-  - intent: countIntent
-  - action: utter_agree
-  - intent: countNumber
-  - intent: countNumber
-  - intent: countNumber
-  - action: utter_thanks
-rules:
-- rule: Say goodbye anytime the user says goodbye
-  steps:
-  - intent: goodbye
-  - action: utter_goodbye
